Issue #292 - use long instead of int on error message for min/max values

+++ WFPMapping/device/survey/src/com/gallatinsystems/survey/device/domain/ValidationRule.java
@@ -157,7 +157,7 @@ public class ValidationRule {
 			if (allowDecimal) {
 				val = num.toString();
 			} else {
-				val = "" + num.intValue();
+				val = "" + num.longValue();
 			}
 		}
 		return val;
